pyproject.toml
[tool.poetry]
name = "gencDNA"
version = "0.1.0"
description = "gencDNA library toolkit"
authors = ["igor-sb <84479009+igor-sb@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mypy = "^0.991"
pytest = "^7.4.0"
pytest-snapshot = "^0.9.0"
biopython = "^1.81"
types-pyyaml = "^6.0.12.11"
pandas = "^2.0.3"
cutadapt = "4.4"
fire = "^0.5.0"
sphinx-rtd-theme = "^1.2.2"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
types-beautifulsoup4 = "^4.12.0.6"
types-requests = "^2.31.0.2"
gff3-parser = "^0.0.5"
tqdm = "^4.66.1"
sniffles = "^2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
autopep8 = "1.6.0"
wemake-python-styleguide = "0.16.1"
unify = "^0.5"
isort = "^5.10.1"
flake8 = "4.0.1"
pre-commit = "^2.20.0"
coverage = "^6.5.0"
ipykernel = "^6.25.0"
prefect = "^2.11.3"
sphinx = "^7.1.2"
pygments = "^2.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
